+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 0.1.101
+
+### Patch Changes
+
+-   Updated dependencies [45fe64e]
+    -   @sap-ux/project-access@1.32.12
+    -   @sap-ux/abap-deploy-config-writer@0.2.37
+    -   @sap-ux/fiori-generator-shared@0.13.37
+    -   @sap-ux/abap-deploy-config-inquirer@1.6.52
+    -   @sap-ux/deploy-config-generator-shared@0.1.56
+
 ## 0.1.100
 
 ### Patch Changes