+++ packages/repo-app-import-sub-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/repo-app-download-sub-generator
 
+## 0.3.167
+
+### Patch Changes
+
+-   Updated dependencies [45fe64e]
+    -   @sap-ux/project-access@1.32.12
+    -   @sap-ux/abap-deploy-config-writer@0.2.37
+    -   @sap-ux/axios-extension@1.24.3
+    -   @sap-ux/fiori-generator-shared@0.13.37
+    -   @sap-ux/launch-config@0.10.39
+    -   @sap-ux/odata-service-inquirer@2.11.13
+    -   @sap-ux/project-input-validator@0.6.34
+    -   @sap-ux/system-access@0.6.29
+    -   @sap-ux/fiori-elements-writer@2.8.11
+    -   @sap-ux/inquirer-common@0.9.9
+
+## 0.3.166
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@2.8.10
+-   @sap-ux/odata-service-inquirer@2.11.12
+
 ## 0.3.165
 
 ### Patch Changes